Run `checkPBsmethod.ipynb` with v0.8.2.dev+3a43c220
In this version, CF of "Water, unspecified natural origin" (natural resource, in water) is set to zero. By doing this, the comparison results become the same in bw and in simapro.
But should this CF be zero? Probably not. So the simapro version of the method has to be corrected...