Add description to the branch
The CBI PB (especially the GW component of it) was taken direcly from simapro version of the method [as opposed to using bw-recipe-2016 package].
This made me realize that GW from bw-recipe-2016 was not including subcategories, which was corrected in this branch (ad685330cb3e).

WIth these changes the `checkPBsmethod.ipynb` generally shows good agreement of CBI PB in bw with the simapro version the method.